Set variables in new location
Just realized the script would attempt to set the HOSTU and HOSTB
variables at the beginning.  If the files were not already setup, this
caused file not found errors to be displayed, then the files should be
setup by the script so it wouldn't be seen again and was a non-issue. So
just moved the variable declaration to be after the check for the files.