output clarification
Found myself confused every now and then with the output of the status
being shown under the file hashes.  To make it more clear I just threw
the hash check/print into a function and call it before and after the
file swap occurs.  Also threw the status into a line to separate and
make it easier to read quickly.

Figured I'd throw the banner into a function also.  Probably not
necessary at all, but in the case I want to display it again, it's all
setup.  Saves me those few seconds of making the function later on...

Lastly removed the 'clear' that was placed right before the files setup
output.  Didn't make sense to have it as it made any output above it
unreadable since everything is completed so quickly.