softlinking hosts.block
Due to the size of the hosts.block file and limited system space, having
two copies (hosts.block & hosts) of the block file in /etc is no longer
possible.  The workaround was as easy as creating a softlink to the
large 'block' file on the sdcard.

This actually works out better when updating the hosts.block file.
Instead of having to remount /system as rewritable and copy the new file
to /etc, it can now just be replaced in the Hosts_swap directory on the
sdcard and the script can continue to be used as normal.